requirements.txt
# Core deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and NLP
transformers>=4.30.0
tokenizers>=0.13.0
sentencepiece>=0.1.99

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
audioread>=3.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine learning utilities
scikit-learn>=1.3.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0

# Logging and experiment tracking (optional)
tensorboard>=2.13.0
# wandb>=0.15.0  # Uncomment if using Weights & Biases

# Development tools (optional)
jupyter>=1.0.0
ipython>=8.12.0

faster-whisper #pentru traducere
nvidia-cublas-cu12 nvidia-cudnn-cu12
peft #multi modal lora